Add component build flags & types file specifier
This should now build the egg with `-O3` and `-d0` flags when installed
via chicken-install.

And yes: I realize this may actually mean that `assume` calls get
compiled out whenever you use the egg, but that seems like a failing of
the CHICKEN build system that you cannot specify debug and release
builds of everything independently.