Refactor predicate tests to use test egg
I also added some tests and removed references to "common" ranges that
are used throughout the test. This has the benefit of:

1. Not relying on global state.
2. The tests won't break if someone is doing fancy stuff with set! in
any of the tests.

Additionally, each test using the test egg now has a much clearer
plain-English description of what is being tested. This does not output
Scheme code to the console, but if you really want to read Scheme then
IMO you should read the test file itself. I tried to be as descriptive
as I could while keeping the names brief.